实现评论功能，第一次用ajax

+++ src/main/java/com/july/community/controller/PublishController.java
@@ -30,7 +30,7 @@ public class PublishController {
             @RequestParam("title") String title,
             @RequestParam("description") String description,
             @RequestParam("tag") String tag,
-            @RequestParam("id") Integer id,
+            @RequestParam("id") Long id,
             HttpServletRequest request,
             Model model
     ){
@@ -60,7 +60,7 @@ public class PublishController {
     }
 
     @GetMapping("publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model){
         QuestionDTO question = questionService.getListById(id);
         model.addAttribute("title",question.getTitle());
